Moved initializeFacebookSDK earlier. ContactsImportedPreview into single component.
