+++ src/App.jsx
@@ -16,6 +16,7 @@ import Header from './js/components/Navigation/Header';
 import HeaderBarSuspense from './js/components/Navigation/HeaderBarSuspense';
 import webAppConfig from './js/config';
 import AppObservableStore, { messageService } from './js/stores/AppObservableStore';
+import initializeFacebookSDK from './js/utils/initializeFacebookSDK';
 import initializejQuery from './js/utils/initializejQuery';
 import RouterV5SendMatch from './js/utils/RouterV5SendMatch';
 // importRemoveCordovaListenersToken1  -- Do not remove this line!
@@ -149,6 +150,10 @@ class App extends Component {
         }, 3000);
       }
     }
+    setTimeout(() => {
+      // We need to start this initialization early since there is a delay getting the FB object in place
+      initializeFacebookSDK();
+    }, 4000);
   }
 
   componentDidCatch (error, info) {
